fix checkstyle
